backend/requirements.txt
Django==3.2.18
djangorestframework==3.14.0
djangorestframework-simplejwt==5.3.0
django-cors-headers==4.1.0
Pillow==9.5.0
python-decouple==3.8
